allow plugins to list / read runtime files
also renamed most of the new functions to be
more specific about what kind of files this is for.
